Move inbound translation from RoutingService to ConversationService.SendMessage
TranslateInboundMessage was called from both InstructDirect and InstructLoop,
leaving the two routing paths to stay in sync by hand. SendMessage is the only
caller of either, so run it once there, ahead of the conversation hook loop.

Running before the hooks means they now evaluate English content, which matters
for RoutingConversationHook: it embeds message.Content and classifies intent
against an English-trained model.

It also means a hook can overwrite Content after SecondaryContent has been set
to the user's original text. Consumers render SecondaryContent in preference to
Content, so a rate-limited or intent-templated reply would be replaced by the
user's own message on its way back to them. Clear SecondaryContent when a hook
rewrites Content, gated on whether translation actually ran so we only ever
clear a value we set ourselves. This matches RoleDialogModel.From, which already
drops SecondaryContent when it replaces Content.
