+++ src/Infrastructure/BotSharp.Core/Conversations/Services/ConversationService.Translation.cs
@@ -1,26 +1,29 @@
 using BotSharp.Abstraction.Infrastructures.Enums;
 
-namespace BotSharp.Core.Routing;
+namespace BotSharp.Core.Conversations.Services;
 
-public partial class RoutingService
+public partial class ConversationService
 {
     private const string TranslationPromptName = "translation_prompt";
 
     /// <summary>
-    /// Normalize an inbound user message to English so routing rules, agent instructions and
-    /// function arguments are always evaluated in English. The user's original text is kept in
-    /// SecondaryContent. Shared by InstructLoop and InstructDirect so every entry point into the
-    /// routing service behaves the same way.
+    /// Normalize an inbound user message to English so conversation hooks, routing rules, agent
+    /// instructions and function arguments are always evaluated in English. The user's original
+    /// text is kept in SecondaryContent. Runs once in SendMessage ahead of the hook loop, so both
+    /// routing paths and every hook observe the same normalized message.
     /// </summary>
-    private async Task TranslateInboundMessage(Agent agent, RoleDialogModel message)
+    /// <returns>
+    /// True when the message was translated and SecondaryContent now holds the user's original text.
+    /// Callers rely on this to tell apart a SecondaryContent this method authored from one it never
+    /// touched, so it must stay false on every early return.
+    /// </returns>
+    private async Task<bool> TranslateInboundMessage(Agent agent, RoleDialogModel message)
     {
         var agentSettings = _services.GetRequiredService<AgentSettings>();
         if (!agentSettings.EnableTranslator)
         {
-            return;
+            return false;
         }
-
-        var states = _services.GetRequiredService<IConversationStateService>();
 
         // The caller supplies the language through the request states; the server does not detect it.
         // TranslationService back-fills StateConst.LANGUAGE only when the state is absent, which cannot
@@ -28,15 +31,15 @@ public partial class RoutingService
         // the /translate endpoint instead.
         // Unknown is excluded to stay in sync with TranslationResponseHook: it means the language has
         // not been resolved, so translating would only paraphrase the user's own words.
-        var language = states.GetState(StateConst.LANGUAGE, LanguageType.ENGLISH);
+        var language = _state.GetState(StateConst.LANGUAGE, LanguageType.ENGLISH);
         if (language == LanguageType.ENGLISH || language == LanguageType.UNKNOWN)
         {
-            return;
+            return false;
         }
 
         // TranslationService reads the prompt template off the agent it is handed, and only the
         // AI Assistant defines it. Fall back to that agent when the executing one has no template,
-        // which is the normal case for the task agents reaching us through InstructDirect.
+        // which is the normal case for the task agents handled by InstructDirect.
         var host = agent;
         if (host?.Templates?.Any(x => x.Name == TranslationPromptName) != true)
         {
@@ -49,5 +52,7 @@ public partial class RoutingService
         message.Content = await translator.Translate(host, message.MessageId, message.Content,
             language: LanguageType.ENGLISH,
             clone: false);
+
+        return true;
     }
 }
